requirements.txt
streamlit==1.38.0
pdfplumber==0.11.4
pillow==10.4.0